refactor: move impact contracts to output crate
