+++ crates/output/src/lib.rs
@@ -32,6 +32,7 @@ mod health_scores;
 mod health_targets;
 mod health_trends;
 mod health_vital_signs;
+mod impact;
 mod inspect_envelopes;
 mod issue_contract;
 mod list_envelopes;
@@ -123,6 +124,11 @@ pub use health_trends::{HealthTrend, TrendCount, TrendDirection, TrendMetric, Tr
 pub use health_vital_signs::{
     RenderFanInTopComponent, RiskProfile, SNAPSHOT_SCHEMA_VERSION, VitalSigns, VitalSignsCounts,
     VitalSignsSnapshot,
+};
+pub use impact::{
+    ContainmentEvent, CrossRepoImpactReport, CrossRepoImpactSchemaVersion, CrossRepoProjectEntry,
+    CrossRepoTotals, EnabledSource, ImpactCounts, ImpactReport, ImpactReportSchemaVersion,
+    ImpactTrendDirection, ResolutionEvent, TrendSummary,
 };
 pub use inspect_envelopes::{
     ExplainOutput, InspectEvidence, InspectEvidenceScope, InspectEvidenceSection,
